Add optional keyword argument to control whether event is included in params
The idea here is that sometimes you don't want delegates to have to take
arguments beyond that which are required. If we know that we are passing,
explicitly, all the arguments that we need, we can strip the event off and not
require the called delegates to take extra **kwargs.

This defaults to True so that we do not violate any existing contracts with
callers of the library.

The reason the option is provided on the route_request method, in lieu of the of
the router instance __init__, is so that it can be controlled per request route.
It may be that you want to include the event for some methods but not others.